+++ includes/avatar-privacy/avatar-handlers/class-default-icons-handler.php
@@ -28,7 +28,8 @@ namespace Avatar_Privacy\Avatar_Handlers;
 
 use Avatar_Privacy\Core;
 use Avatar_Privacy\Default_Icons;
-use Avatar_Privacy\Image_Tools;
+
+use Avatar_Privacy\Tools\Images as Image_Tools;
 
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
