+++ includes/avatar-privacy/default-icons/generator/class-wavatar.php
@@ -27,7 +27,7 @@
 
 namespace Avatar_Privacy\Default_Icons\Generator;
 
-use Avatar_Privacy\Image_Tools;
+use Avatar_Privacy\Tools\Images as Image_Tools;
 
 /**
  * A wavatar generator.
