+++ includes/avatar-privacy/tools/images/class-image-stream.php
@@ -24,7 +24,7 @@
  * @license http://www.gnu.org/licenses/gpl-2.0.html
  */
 
-namespace Avatar_Privacy\Image_Tools;
+namespace Avatar_Privacy\Tools\Images;
 
 /**
  * A persistent memory stream for storing temporary images.
