+++ includes/avatar-privacy/upload-handlers/class-upload-handler.php
@@ -27,9 +27,10 @@
 namespace Avatar_Privacy\Upload_Handlers;
 
 use Avatar_Privacy\Core;
-use Avatar_Privacy\Image_Tools;
 
 use Avatar_Privacy\Data_Storage\Filesystem_Cache;
+
+use Avatar_Privacy\Tools\Images as Image_Tools;
 
 /**
  * Handles image uploads.
